+++ packages/backend/src/api/resources/Deserializer.ts
@@ -15,10 +15,10 @@ import {
   InstanceSettings,
   Invitation,
   JwtTemplate,
+  M2MToken,
   Machine,
   MachineScope,
   MachineSecretKey,
-  MachineToken,
   OauthAccessToken,
   OAuthApplication,
   Organization,
@@ -37,6 +37,8 @@ import {
   User,
 } from '.';
 import { AccountlessApplication } from './AccountlessApplication';
+import { CommercePlan } from './CommercePlan';
+import { Feature } from './Feature';
 import type { PaginatedResponseJSON } from './JSON';
 import { ObjectType } from './JSON';
 import { WaitlistEntry } from './WaitlistEntry';
@@ -141,8 +143,8 @@ function jsonToObject(item: any): any {
       return MachineScope.fromJSON(item);
     case ObjectType.MachineSecretKey:
       return MachineSecretKey.fromJSON(item);
-    case ObjectType.MachineToken:
-      return MachineToken.fromJSON(item);
+    case ObjectType.M2MToken:
+      return M2MToken.fromJSON(item);
     case ObjectType.OauthAccessToken:
       return OauthAccessToken.fromJSON(item);
     case ObjectType.OAuthApplication:
@@ -179,6 +181,10 @@ function jsonToObject(item: any): any {
       return User.fromJSON(item);
     case ObjectType.WaitlistEntry:
       return WaitlistEntry.fromJSON(item);
+    case ObjectType.CommercePlan:
+      return CommercePlan.fromJSON(item);
+    case ObjectType.Feature:
+      return Feature.fromJSON(item);
     default:
       return item;
   }